requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.4.2
sqlalchemy==2.0.23
openai==1.3.0
httpx==0.24.1
--find-links https://download.pytorch.org/whl/torch_stable.html
numpy>=1.26.0
spacy==3.7.2
python-multipart==0.0.6
tiktoken==0.5.2
pandas==2.0.3
tqdm==4.66.1
fastapi-jwt-auth==0.5.0
scikit-learn==1.3.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
streamlit==1.32.0
emoji==2.10.1